+++ packages/core/src/types/RenderPage.ts
@@ -14,6 +14,8 @@ import type { Slot } from './Slot';
 export interface RenderPageProps {
     annotationLayer: Slot;
     canvasLayer: Slot;
+    // Is the canvas layer rendered completely?
+    canvasLayerRendered: boolean;
     doc: PdfJs.PdfDocument;
     height: number;
     pageIndex: number;
@@ -21,7 +23,11 @@ export interface RenderPageProps {
     scale: number;
     svgLayer: Slot;
     textLayer: Slot;
+    // Is the text layer rendered completely?
+    textLayerRendered: boolean;
     width: number;
+    // Mark as the page rendered completely
+    markRendered(pageIndex: number): void;
 }
 
 export type RenderPage = (props: RenderPageProps) => React.ReactElement;